feat: 04 SpringApplication으로 최종 구현! (이후 SpringApplication을 사용할 것임)
- 지금까지 구현한 main()의 코드를 MySpringApplication 클래스를 만들어 run() 메소드로 넣고, 메인 클래스를 파라미터로 받아서 사용하도록 만들면, 스프링 부트의 main() 메소드가 있는 클래스와 유사한 코드가 만들어진다.
